FISH-317 Change XMPP Notifier Package
