+++ appserver/payara-appserver-modules/notification-xmpp-core/src/main/java/fish/payara/extras/notifier/xmpp/XmppNotifier.java
@@ -37,7 +37,7 @@
  * only if the new code is made subject to such option by the copyright
  * holder.
  */
-package fish.payara.notification.xmpp;
+package fish.payara.extras.notifier.xmpp;
 
 import java.io.IOException;
 import java.util.logging.Level;
